+++ src/cpp/fastdds/builtin/type_lookup_service/TypeLookupRequestListener.cpp
@@ -24,6 +24,7 @@
 #include <fastdds/builtin/type_lookup_service/TypeLookupManager.hpp>
 #include <rtps/participant/RTPSParticipantImpl.h>
 #include <rtps/RTPSDomainImpl.hpp>
+#include <rtps/reader/StatefulReader.hpp>
 
 using eprosima::fastrtps::rtps::RTPSReader;
 using eprosima::fastrtps::rtps::CacheChange_t;
