+++ src/cpp/rtps/builtin/discovery/endpoint/EDPServer.cpp
@@ -24,7 +24,7 @@
 #include <fastdds/rtps/attributes/ReaderAttributes.h>
 #include <fastdds/rtps/history/WriterHistory.h>
 #include <fastdds/rtps/writer/StatefulWriter.h>
-#include <fastdds/rtps/reader/StatefulReader.h>
+#include <rtps/reader/StatefulReader.hpp>
 
 using namespace ::eprosima::fastrtps::rtps;
