+++ src/cpp/rtps/builtin/discovery/participant/PDP.cpp
@@ -33,8 +33,6 @@
 #include <fastdds/rtps/history/ReaderHistory.h>
 #include <fastdds/rtps/history/WriterHistory.h>
 #include <fastdds/rtps/participant/RTPSParticipantListener.h>
-#include <fastdds/rtps/reader/StatefulReader.h>
-#include <fastdds/rtps/reader/StatelessReader.h>
 #include <fastdds/rtps/writer/StatelessWriter.h>
 
 #include <fastdds/builtin/type_lookup_service/TypeLookupManager.hpp>
@@ -50,6 +48,8 @@
 #include <rtps/network/utils/external_locators.hpp>
 #include <rtps/resources/TimedEvent.h>
 #include <rtps/participant/RTPSParticipantImpl.h>
+#include <rtps/reader/StatefulReader.hpp>
+#include <rtps/reader/StatelessReader.hpp>
 #include <utils/shared_mutex.hpp>
 #include <utils/TimeConversion.hpp>
