+++ src/cpp/rtps/builtin/discovery/participant/PDPSimple.cpp
@@ -27,8 +27,6 @@
 #include <fastdds/rtps/history/ReaderHistory.h>
 #include <fastdds/rtps/history/WriterHistory.h>
 #include <fastdds/rtps/participant/RTPSParticipantListener.h>
-#include <fastdds/rtps/reader/StatefulReader.h>
-#include <fastdds/rtps/reader/StatelessReader.h>
 #include <fastdds/rtps/writer/StatelessWriter.h>
 
 #include <fastdds/builtin/type_lookup_service/TypeLookupManager.hpp>
@@ -44,6 +42,8 @@
 #include <rtps/builtin/liveliness/WLP.h>
 #include <rtps/history/TopicPayloadPoolRegistry.hpp>
 #include <rtps/participant/RTPSParticipantImpl.h>
+#include <rtps/reader/StatefulReader.hpp>
+#include <rtps/reader/StatelessReader.hpp>
 #include <rtps/resources/TimedEvent.h>
 
 namespace eprosima {
