+++ src/cpp/rtps/builtin/discovery/participant/simple/SimplePDPEndpoints.hpp
@@ -22,13 +22,12 @@
 #include <memory>
 
 #include <fastdds/rtps/builtin/data/BuiltinEndpoints.hpp>
-#include <fastdds/rtps/reader/StatelessReader.h>
 #include <fastdds/rtps/writer/StatelessWriter.h>
-
 #include <rtps/builtin/BuiltinReader.hpp>
 #include <rtps/builtin/BuiltinWriter.hpp>
 #include <rtps/builtin/discovery/participant/PDPEndpoints.hpp>
 #include <rtps/history/ITopicPayloadPool.h>
+#include <rtps/reader/StatelessReader.hpp>
 
 namespace eprosima {
 namespace fastdds {
