+++ src/cpp/rtps/builtin/liveliness/WLP.cpp
@@ -27,7 +27,6 @@
 #include <fastdds/rtps/builtin/data/WriterProxyData.h>
 #include <fastdds/rtps/history/ReaderHistory.h>
 #include <fastdds/rtps/history/WriterHistory.h>
-#include <fastdds/rtps/reader/StatefulReader.h>
 #include <fastdds/rtps/resources/ResourceEvent.h>
 
 #include <fastdds/rtps/writer/StatefulWriter.h>
@@ -38,6 +37,7 @@
 #include <rtps/builtin/liveliness/WLPListener.h>
 #include <rtps/history/TopicPayloadPoolRegistry.hpp>
 #include <rtps/participant/RTPSParticipantImpl.h>
+#include <rtps/reader/StatefulReader.hpp>
 #include <rtps/resources/TimedEvent.h>
 #include <rtps/writer/LivelinessManager.hpp>
 #include <utils/TimeConversion.hpp>
