+++ src/cpp/rtps/reader/StatefulReader.cpp
@@ -17,6 +17,8 @@
  *
  */
 
+#include <rtps/reader/StatefulReader.hpp>
+
 #include <cassert>
 #include <mutex>
 #include <thread>
@@ -25,7 +27,6 @@
 #include <fastdds/rtps/common/VendorId_t.hpp>
 #include <fastdds/rtps/history/ReaderHistory.h>
 #include <fastdds/rtps/reader/ReaderListener.h>
-#include <fastdds/rtps/reader/StatefulReader.h>
 
 #include <rtps/builtin/BuiltinProtocols.h>
 #include <rtps/builtin/liveliness/WLP.h>
