+++ test/unittest/rtps/reader/WriterProxyStopTest.cpp
@@ -19,9 +19,9 @@
 
 #include <fastdds/rtps/builtin/data/WriterProxyData.h>
 #include <fastdds/rtps/reader/RTPSReader.h>
-#include <fastdds/rtps/reader/StatefulReader.h>
 #include <fastdds/rtps/resources/ResourceEvent.h>
 #include <rtps/participant/RTPSParticipantImpl.h>
+#include <rtps/reader/StatefulReader.hpp>
 #include <rtps/reader/WriterProxy.h>
 #include <rtps/reader/WriterProxy.cpp>
 #include <rtps/resources/TimedEvent.h>
